requirements.txt
langgraph
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
fastapi
uvicorn
pydantic
python-dotenv
numpy